+++ StarkMine-Game/Assets/_Project/_Scripts/Game/_UI/UpgradeSpaceStationUI.cs
@@ -57,7 +57,7 @@ public class UpgradeSpaceStationUI : BasePopup
     private void OnEnable()
     {
         DataManager.Instance.OnMineCoinUpdate += DataManagerOnOnMineCoinUpdate;
-        int mineCoins = DataManager.Instance.MineCoin;
+        double mineCoins = DataManager.Instance.MineCoin;
         bool isEnoughCoin = mineCoins > _stationData.GetCostForNextLevel();
         balanceText.text = Helpers.FormatCurrencyNumber(mineCoins) + " $MINE";
         notificationText.SetActive(!isEnoughCoin);
@@ -71,7 +71,7 @@ public class UpgradeSpaceStationUI : BasePopup
 
     private void DataManagerOnOnMineCoinUpdate(object sender, DataManager.OnMineCoinUpdateEventArgs e)
     {
-        int mineCoins = e.NewMineCoin;
+        double mineCoins = e.NewMineCoin;
         bool isEnoughCoin = mineCoins > _stationData.GetCostForNextLevel();
         balanceText.text = Helpers.FormatCurrencyNumber(mineCoins) + " $MINE";
         notificationText.SetActive(!isEnoughCoin);
